Move the error msgs near the buttons so that it's not buried inside the scrollable area
